Divided offline api to separate independent module
