+++ inference-engine/ie_bridges/python/src/openvino/inference_engine/ie_api.pyx
@@ -56,9 +56,6 @@ cdef c_map_to_dict(map[string, string] c_map):
 def get_version():
     return C.get_version().decode()
 
-def ApplyMOCTransformations(IENetwork network, bool cf):
-    C.ApplyMOCTransformations(network.impl, cf)
-
 ## This class defines Tensor description
 cdef class TensorDesc:
     def __eq__(self, other : TensorDesc):
