+++ inference-engine/ie_bridges/python/src/openvino/inference_engine/ie_api_impl.cpp
@@ -2,8 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include <moc_transformations.hpp>
-#include <ngraph/pass/manager.hpp>
 #include "ie_api_impl.hpp"
 #include "hetero/hetero_plugin_config.hpp"
 #include "ie_iinfer_request.hpp"
@@ -614,10 +612,4 @@ PyObject *InferenceEnginePython::IECore::getMetric(const std::string &deviceName
 PyObject *InferenceEnginePython::IECore::getConfig(const std::string &deviceName, const std::string &name) {
     InferenceEngine::Parameter param = actual.GetConfig(deviceName, name);
     return parse_parameter(param);
-}
-
-void InferenceEnginePython::ApplyMOCTransformations(InferenceEnginePython::IENetwork network, bool cf) {
-    ngraph::pass::Manager manager;
-    manager.register_pass<MOCTransformations>(cf);
-    manager.run_passes(network.actual->getFunction());
 }
