+++ inference-engine/ie_bridges/python/src/openvino/inference_engine/ie_api_impl.hpp
@@ -184,6 +184,4 @@ std::unique_ptr<T> make_unique(Args &&... args) {
 }
 
 std::string get_version();
-
-void ApplyMOCTransformations(InferenceEnginePython::IENetwork network, bool cf);
 };  // namespace InferenceEnginePython
