+++ inference-engine/ie_bridges/python/src/openvino/inference_engine/offline_api.pyx
@@ -1,0 +1,22 @@
+"""
+ Copyright (C) 2020 Intel Corporation
+
+ Licensed under the Apache License, Version 2.0 (the 'License');
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an 'AS IS' BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+"""
+from .cimport offline_api_impl_defs as C
+from .ie_api cimport IENetwork
+
+from libcpp cimport bool
+
+def ApplyMOCTransformations(IENetwork network, bool cf):
+    C.ApplyMOCTransformations(network.impl, cf)
