+++ inference-engine/ie_bridges/python/src/openvino/inference_engine/offline_api_impl.hpp
@@ -1,0 +1,14 @@
+// Copyright (C) 2018-2020 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#pragma once
+
+#include "Python.h"
+#include "ie_api_impl.hpp"
+
+namespace InferenceEnginePython {
+
+void ApplyMOCTransformations(InferenceEnginePython::IENetwork network, bool cf);
+
+};  // namespace InferenceEnginePython
