+++ inference-engine/ie_bridges/python/src/openvino/inference_engine/offline_api_impl_defs.pxd
@@ -1,0 +1,5 @@
+from libcpp cimport bool
+from .ie_api_impl_defs cimport IENetwork
+
+cdef extern from "offline_api_impl.hpp" namespace "InferenceEnginePython":
+    cdef void ApplyMOCTransformations(IENetwork network, bool cf)
